+++ src/Handyman.AspNetCore/src/ETags/EnsureRequestHasIfMatchHeader.cs
@@ -1,0 +1,4 @@
+﻿namespace Handyman.AspNetCore.ETags
+{
+    internal class EnsureRequestHasIfMatchHeader { }
+}