+++ src/Handyman.AspNetCore/src/ETags/IETagConverter.cs
@@ -1,8 +1,7 @@
 ﻿namespace Handyman.AspNetCore.ETags
 {
-    public interface IETagConverter
+    internal interface IETagConverter
     {
         string FromByteArray(byte[] bytes);
-        byte[] ToByteArray(string eTag);
     }
 }