+++ src/Handyman.AspNetCore/src/ETags/ModelBinding/ETagModelBinder.cs
@@ -1,36 +1,17 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
+﻿using Microsoft.AspNetCore.Mvc.ModelBinding;
+using System.Threading.Tasks;
 
 namespace Handyman.AspNetCore.ETags.ModelBinding
 {
     internal class ETagModelBinder : IModelBinder
     {
-        private readonly IETagValidator _validator;
-
-        public ETagModelBinder(IETagValidator validator)
-        {
-            _validator = validator;
-        }
-
         public Task BindModelAsync(ModelBindingContext bindingContext)
         {
             var headers = bindingContext.HttpContext.Request.Headers;
 
             if (headers.TryGetValue(bindingContext.BinderModelName, out var values))
             {
-                var eTag = values.ToString();
-
-                if (_validator.IsValidETag(eTag))
-                {
-                    bindingContext.Result = ModelBindingResult.Success(eTag);
-                }
-                else
-                {
-                    var key = bindingContext.ModelName;
-                    var error = "Invalid ETag format.";
-                    bindingContext.ModelState.AddModelError(key, error);
-                    bindingContext.Result = ModelBindingResult.Failed();
-                }
+                bindingContext.Result = ModelBindingResult.Success(values.ToString());
             }
 
             return Task.CompletedTask;