+++ src/Handyman.AspNetCore/src/ETags/PreconditionFailedException.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Handyman.AspNetCore.ETags
+{
+    internal class PreconditionFailedException : Exception { }
+}